requirements.txt
# Web Framework & Server
fastapi
uvicorn
python-multipart

# Database & ORM
sqlalchemy
psycopg2

# Auth & Security
python-jose
pwdlib[argon2]

# Validation & Settings
pydantic
pydantic-settings
pydantic[email]

# Environment
python-dotenv

# Agent & Orchestration (LangChain / LangGraph)
langchain
langchain-core
langchain-community
langchain-text-splitters
langgraph
langchainhub
langsmith

# LLMs & Inference integrations
langchain-groq
langchain-google-genai

# Embeddings & Vector Stores
faiss-cpu
PyPDF2
pypdf
